+++ src/main/java/com/example/testypie/domain/user/controller/UserController.java
@@ -1,23 +1,21 @@
 package com.example.testypie.domain.user.controller;
 
 
-import com.example.testypie.global.jwt.JwtUtil;
-import com.example.testypie.domain.feedback.entity.Feedback;
-import com.example.testypie.domain.user.dto.*;
-import com.example.testypie.domain.user.entity.User;
-import com.example.testypie.domain.user.service.UserInfoService;
+import com.example.testypie.domain.user.dto.LoginRequestDTO;
+import com.example.testypie.domain.user.dto.MessageDTO;
+import com.example.testypie.domain.user.dto.SignUpRequestDTO;
 import com.example.testypie.domain.user.service.UserService;
-import com.example.testypie.global.security.UserDetailsImpl;
+import com.example.testypie.global.jwt.JwtUtil;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
-import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
-import java.util.List;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
 @RestController
@@ -26,8 +24,6 @@ import java.util.List;
 public class UserController {
 
     private final UserService userService;
-
-    private final UserInfoService userInfoService;
 
     private final JwtUtil jwtUtil;
 
@@ -44,116 +40,9 @@ public class UserController {
     //로그인
     @PostMapping("/login")
     public ResponseEntity<MessageDTO> login(@RequestBody LoginRequestDTO req,
-            HttpServletResponse res) {
+                                            HttpServletResponse res) {
         userService.login(req);
         res.setHeader(JwtUtil.AUTHORIZATION_HEADER, jwtUtil.createToken(req.account()));
         return ResponseEntity.ok().body(new MessageDTO("로그인 성공", HttpStatus.OK.value()));
-    }
-
-    //프로필 조회
-    @GetMapping("/{account}")
-    public ResponseEntity<ProfileResponseDTO> getProfile(@PathVariable String account) {
-        User user = userInfoService.findProfile(account);
-        return ResponseEntity.ok().body(ProfileResponseDTO.of(user));
-    }
-
-    //프로필 수정
-    @PatchMapping("/{account}")
-    public ResponseEntity<?> updateProfile(@PathVariable String account,
-            @RequestBody ProfileRequestDTO req) {
-
-        try {
-            ProfileResponseDTO res = userInfoService.updateProfile(account, req);
-            return ResponseEntity.ok(res);
-        } catch (NoSuchElementException e) {
-            return ResponseEntity.notFound().build();
-        } catch (Exception e) {
-            return ResponseEntity.internalServerError()
-                    .body(new MessageDTO("업데이트에 실패했습니다.", HttpStatus.BAD_REQUEST.value()));
-        }
-    }
-
-    // 2024-01-12
-    // product 등록 이력 조회
-    // user가 자신의 프로필에서 product 등록이력을 조회하는 서비스입니다.
-    // 로직은 아래와 같습니다.
-    // 1. 유효한 사용자인지 체크합니다.
-    // 2. 유효한 사용자일 경우 등록한 product를 전부 가져옵니다.
-    // 3. RegisteredProductResponstDTO객체는 productId, product title, createdAt, closedAt를 가집니다.
-    @GetMapping("/{account}/registeredProducts")
-    public ResponseEntity<List<RegisteredProductResponseDTO>> getRegisteredProducts(@PathVariable String account, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        // 1.
-        userInfoService.checkSameUser(account, userDetails.getUsername());
-
-        // 2, 3.
-        List<RegisteredProductResponseDTO> res = userInfoService.getUserProducts(account);
-
-        return ResponseEntity.ok().body(res);
-    }
-
-    // 2024-01-15
-    // product 참여 이력 조회
-    // user가 자신의 프로필에서 product 등록 이력을 조회하는 서비스입니다.
-    // 로직은 아래와 같습니다.
-    // 1. 유효한 사용자인지 체크합니다.
-    // 2. 해당 유저가 작성한 feedback을 모두 조회합니다.
-    // 3. 조회한 feedback들에서 product이름, **feedback별점(미구현), feedback 작성일시를 가져옵니다.
-    @GetMapping("{account}/participatedProducts")
-    public ResponseEntity<List<ParticipatedProductResponseDTO>> getParticipatedProducts(@PathVariable String account, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        // 1.
-        userInfoService.checkSameUser(account, userDetails.getUsername());
-
-        // 2. 3.
-        List<ParticipatedProductResponseDTO> res = userInfoService.getUserFeedbacks(account);
-
-        return ResponseEntity.ok().body(res);
-    }
-
-
-    // 2024-01-17
-    // 테스트 유저가 지금 까지 받은 점수를 평균내서 보여주는 기능입니다.
-    // 1. 유효한 사용자인지 확인합니다.
-    // 2. 자신이 작성한 모든 feedback을 조회한후 rating의 평균점수를 계산해옵니다.
-    // 3. 결과는 AverageRatingResponseDTO에 담겨 보내집니다.
-    @GetMapping("{account}/averageStarRating")
-    public ResponseEntity<AverageRatingResponseDTO> getAverageStarRating(@PathVariable String account,
-                                                  @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        // 1.
-        userInfoService.checkSameUser(account, userDetails.getUsername());
-
-        // 2.
-        double averageRating = userInfoService.getAverageRating(account);
-
-        // 3.
-        AverageRatingResponseDTO res = new AverageRatingResponseDTO(averageRating);
-        return ResponseEntity.ok().body(res);
-    }
-
-    // 2024-01-16
-    // product 등록자가 tester의 feedback에 별점을 매기는 메서드입니다.
-    // 로직은 아래와 같습니다.
-    // 1. product 등록자와 userDetails의 유저가 같은지 확인합니다.
-    // 2. productId와 FeedbackId가 유효한지 확인합니다.
-    // 3. assignRatingStarAfFeedback(Feedback, ratingstarrequestDTO)로 별점(rating)을 매깁니다. 예) 별점 3점 :: assignRating(RatingStar.THREE) -> rating = 3.0;
-    // 4. 별점을 double rating column에 넣습니다.
-    @PostMapping("/{account}/ratingStar/{productId}/{feedbackId}")
-    public ResponseEntity<MessageDTO> assignRatingStarToFeedback(@PathVariable String account, @PathVariable Long productId,
-                                                                 @PathVariable Long feedbackId,
-                                                                 @Valid @RequestBody RatingStarRequestDTO req,
-                                                                 @AuthenticationPrincipal UserDetailsImpl userDetails) {
-
-
-        // 1.
-        userInfoService.checkSameUser(account, userDetails.getUsername());
-
-        // 2.
-        Feedback feedback = userInfoService.getValidFeedback(productId, feedbackId);
-
-        // 3
-        userInfoService.assignRatingStarAtFeedback(feedback, req);
-
-        //4.
-        String message = String.format("별점이 %.1f점 매겨졌습니다.", req.rating());
-        return ResponseEntity.ok().body(new MessageDTO(message, HttpStatus.OK.value()));
     }
 }
