+++ src/main/java/com/example/testypie/domain/user/controller/ViewController.java
@@ -1,8 +1,7 @@
-package com.example.testypie.user.controller;
+package com.example.testypie.domain.user.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
